Response and Request refactored
